requirementx.txt
async-timeout==4.0.3
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
PyMySQL==1.1.0
python-dotenv==1.0.1
redis==5.0.4
requests==2.31.0
six==1.16.0
soupsieve==2.5
spotipy==2.23.0
urllib3==2.2.1
